Reorganize `pillarbox-demo-tv` structure
Move examples

Move lists
