+++ pillarbox-demo-tv/src/main/res/layout/activity_leanback_player.xml
@@ -4,6 +4,6 @@
   -->
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/fragment_container"
-    android:name="ch.srgssr.pillarbox.demo.tv.player.leanback.LeanbackPlayerFragment"
+    android:name="ch.srgssr.pillarbox.demo.tv.ui.player.leanback.LeanbackPlayerFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
